Consolidate ygoprodeck-utils into ygoprodeck-client
