+++ pom.xml
@@ -11,7 +11,6 @@
 
     <modules>
         <module>ygoprodeck-model</module>
-        <module>ygoprodeck-utils</module>
         <module>ygoprodeck-testutils</module>
         <module>ygoprodeck-client</module>
     </modules>
