+++ ygoprodeck-client/src/main/java/io/github/magicdgs/gaming/ygoprodeck/client/YgoprodeckClient.java
@@ -5,7 +5,7 @@ import io.github.magicdgs.gaming.ygoprodeck.api.DatabaseApi;
 import io.github.magicdgs.gaming.ygoprodeck.api.ImagesApi;
 import io.github.magicdgs.gaming.ygoprodeck.client.internal.retrofit.YgoprodeckExceptionCallAdapterFactory;
 import io.github.magicdgs.gaming.ygoprodeck.model.exception.YgoprodeckException;
-import io.github.magicdgs.gaming.ygoprodeck.utils.okhttp.OkhttpClientBuilder;
+import io.github.magicdgs.gaming.ygoprodeck.client.internal.okhttp.OkhttpClientBuilder;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
